# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIXTYSIX API",
    "version": "0.2.0",
    "description": "구현 전 계약. B0 로그인 2종/공개 모집 반영. 서버 시간 권위, unknown request fields 거부, 업무 성공 쓰기만 멱등 캐시. 세부 권장 정책은 BACKEND-B0-SPEC.md 참조. 인증 cookie 실제 이름은 T03 adapter에서 고정. 모든 보호/인증 응답 Cache-Control: no-store."
  },
  "servers": [
    {
      "url": "/",
      "description": "동일 origin /v1 프록시; 배포 미구현"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/v1/auth/code": {
      "post": {
        "operationId": "sendCode",
        "summary": "이메일 로그인 인증번호 요청",
        "tags": [
          "auth"
        ],
        "responses": {
          "202": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "존재 여부와 무관한 202 형태·안내. 이메일은 소문자 정규화하며 IP/이메일/브라우저별 DB rate limit 적용. 6자리·5분·오답 3회. 공급자 수락 후 202, 수락은 실제 배달 보장이 아님. 기존 challenge 재사용 금지. Origin/CSRF 실패는 403.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/auth/verify": {
      "post": {
        "operationId": "verifyCode",
        "summary": "인증번호 검증",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "challenge에 저장된 목적·브라우저·사용자·intent를 사용. reauth/link는 시작한 현재 세션 필수. 로그인 성공은 세션 cookie 설정, reauth는 동일 사용자만 fresh 갱신, link는 서버의 intent 검증 상태만 갱신. code 실패는 계정 존재를 밝히지 않는 INVALID_CHALLENGE.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$"
                  }
                },
                "required": [
                  "challengeId",
                  "code"
                ]
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "현재 세션 폐기",
        "tags": [
          "auth"
        ],
        "responses": {
          "204": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "브라우저 쿠키도 만료. 이미 로그아웃한 문맥도 CSRF 검사 후 204.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "security": []
      }
    },
...
    "/v1/cohorts": {
      "get": {
        "operationId": "listCohorts",
        "summary": "공개 모집 목록",
        "tags": [
          "cohorts"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CohortPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "시작 전 예정/모집 코호트 목록. startsAt ASC,id ASC 커서, 실제 참여 인원 사용. 인증 요청의 canJoin은 개인 슬롯 및 동일 기수 취소 이력을 반영한다. 비회원 canJoin은 모집 조건이며 실제 참여에는 로그인이 필요하다.",
        "parameters": [
          {
            "in": "query",
            "name": "habitId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "security": []
      }
    },
    "/v1/cohorts/{cohortId}": {
      "get": {
        "operationId": "getCohort",
        "summary": "공개 코호트 상세",
        "tags": [
          "cohorts"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cohort"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "cohortId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": []
      }
    },
    "/v1/memberships": {
      "post": {
        "operationId": "joinCohort",
        "summary": "코호트 참여",
        "tags": [
          "memberships"
        ],
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          },
          {
            "$ref": "#/components/parameters/Idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "cohortId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "cohortId"
                ]
              }
            }
          }
        },
        "description": "시작 전, 정원 및 활성/대기 슬롯 1개 조건을 확인한다. 취소한 동일 기수 재참여는 JOIN_CLOSED. 동일 키 성공은 원래 응답을 반환하며 변경된 입력/경로로 키를 재사용하면 IDEMPOTENCY_CONFLICT. 계정별로 네 업무 쓰기가 한도를 공유한다. 새 요청과 업무 거부/키 충돌은 분당 20회·시간당 120회, 성공 응답의 동일 키·내용 재조회는 별도 분당 60회. 한도 초과는 429 RATE_LIMITED와 초 단위 Retry-After를 반환하며 업무 변경은 없다. 정상 재조회는 새 요청 한도를 소모하지 않는다."
      }
    },
    "/v1/memberships/{membershipId}/cancel": {
      "post": {
        "operationId": "cancelMembership",
        "summary": "시작 전 참여 취소",
        "tags": [
          "memberships"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "본인의 시작 전 참여만 취소한다. 취소 사실과 슬롯 해제·감사·멱등 성공을 함께 저장한다. 동일 기수 재참여 불가. 이미 성공한 키는 시작 이후에도 원래 성공을 반환한다. leave 명령은 아직 미구현. 계정별로 네 업무 쓰기가 한도를 공유한다. 새 요청과 업무 거부/키 충돌은 분당 20회·시간당 120회, 성공 응답의 동일 키·내용 재조회는 별도 분당 60회. 한도 초과는 429 RATE_LIMITED와 초 단위 Retry-After를 반환하며 업무 변경은 없다. 정상 재조회는 새 요청 한도를 소모하지 않는다.",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Csrf"
          },
          {
            "$ref": "#/components/parameters/Idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "confirm": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                },
                "required": [
                  "confirm"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-sixtysix.session",
        "description": "HTTPS: __Host-sixtysix.session, Path=/·HttpOnly·Secure·SameSite=Lax·Domain 없음. 로컬 HTTP만 sixtysix.session. DB에는 토큰 SHA-256만 저장, 개발 기본 절대 만료 7일."
      }
    },
    "parameters": {
      "Idempotency": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "같은 동작의 네트워크 재시도는 같은 UUID, 변경된 입력은 새 UUID. method/path/body를 요청 hash에 포함."
      },
      "Csrf": {
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "GET /v1/auth/context 발급값. 브라우저/세션에 결합, Origin 검사와 함께 사용."
      },
      "Cursor": {
        "in": "query",
        "name": "cursor",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "Limit": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50,
          "default": 20
        }
      }
    },
    "responses": {
      "400": {
        "description": "입력 또는 인증 challenge/state 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "401": {
        "description": "유효 세션 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "403": {
        "description": "운영 권한 부족 또는 REAUTH_REQUIRED",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "404": {
        "description": "없거나 접근 불가한 업무 리소스",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "409": {
        "description": "업무/멱등/연결 충돌",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "429": {
        "description": "속도 제한",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "재시도까지 초"
          }
        }
      },
      "503": {
        "description": "일시 장애",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Membership": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cohortId": {
            "type": "string",
            "format": "uuid"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "leftAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "cohortId",
          "joinedAt",
          "cancelledAt",
          "leftAt"
        ]
      },
...
      "Cohort": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "habitId": {
            "type": "string",
            "format": "uuid"
          },
          "generation": {
            "type": "string",
            "minLength": 1
          },
          "policyVersion": {
            "type": "integer",
            "minimum": 1
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "durationDays": {
            "type": "integer",
            "enum": [
              66
            ]
          },
          "capacity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          },
          "minParticipants": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          },
          "participantCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30
          },
          "recruitmentOpensAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "recruiting",
              "active",
              "ended",
              "cancelled"
            ]
          },
          "finalSubmissionAt": {
            "type": "string",
            "format": "date-time"
          },
          "canJoin": {
            "type": "boolean"
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "habitId",
          "generation",
          "policyVersion",
          "startsAt",
          "startDate",
          "durationDays",
          "capacity",
          "minParticipants",
          "participantCount",
          "recruitmentOpensAt",
          "status",
          "finalSubmissionAt",
          "canJoin",
          "cancellationReason"
        ]
      },
      "Available": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "targetDay": {
            "type": "integer",
            "minimum": 1,
            "maximum": 66
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          },
          "late": {
            "type": "boolean"
          },
          "closesAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "ok",
          "targetDay",
          "targetDate",
          "late",
          "closesAt"
        ]
      },
      "Unavailable": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "BEFORE_START",
              "ALREADY_FILLED",
              "ENDED",
              "CANCELLED",
              "LEFT"
            ]
          }
        },
        "required": [
          "ok",
          "reason"
        ]
      },
      "Availability": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/Available"
          },
          {
            "$ref": "#/components/schemas/Unavailable"
          }
        ]
      },
...
      "CodeAccepted": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "challengeId",
          "expiresAt",
          "message"
        ]
      },
...
      "VerifyResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "login",
              "reauth",
              "link"
            ]
          },
          "nextAction": {
            "type": "string",
            "enum": [
              "SIGNED_IN",
              "REAUTHENTICATED",
              "COMPLETE_LINK"
            ]
          },
          "intentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "purpose",
          "nextAction",
          "intentId"
        ]
      },
      "ErrorDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "field": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_REQUEST",
                  "INVALID_TEXT",
                  "INVALID_MEDIA",
                  "INVALID_CHALLENGE",
                  "INVALID_OAUTH_STATE",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "REAUTH_REQUIRED",
                  "NOT_FOUND",
                  "COHORT_FULL",
                  "JOIN_CLOSED",
                  "ACTIVE_MEMBERSHIP_EXISTS",
                  "ALREADY_FILLED",
                  "PASS_EXHAUSTED",
                  "BEFORE_START",
                  "ENDED",
                  "TARGET_CHANGED",
                  "IDEMPOTENCY_CONFLICT",
                  "IDENTITY_ALREADY_LINKED",
                  "ACCOUNT_NOT_LINKED",
                  "LAST_IDENTITY",
                  "INVALID_STATE",
                  "RATE_LIMITED",
                  "TEMPORARILY_UNAVAILABLE",
                  "COHORT_EXISTS"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "requestId": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "$ref": "#/components/schemas/ErrorDetails"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "CohortPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cohort"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
    }
  }
}